Part rename
Renamed the header part to “MAX / Turbot Compatible LSE” to make it
easier to read and to show its compatibility between the MinnowBoard
MAX and MinnowBoard Turbot.  Design files were updated to make use of
the new library with the renamed parts, Schematic.pdf and Gerber files
updated as well.